Organized the frames with functions, added a play button and a play menu; put all the GUI related stuff in app.py, and put the old code in app.py in run.py
